src/requirements.txt
jsondiff==2.0.0
pycryptodome==3.14.1
requests==2.27.1
